Split out AuthorizationMiddleware
This commit splits out the `AuthorizationMiddleware` into its own file
to keep mod.rs clean.
