+++ libsplinter/src/rest_api/auth/actix/mod.rs
@@ -14,187 +14,32 @@
 
 //! Authorization middleware for the Actix REST API
 
+mod middleware;
 mod transform;
 
+pub use middleware::AuthorizationMiddleware;
 pub use transform::Authorization;
-
-use actix_web::dev::*;
-use actix_web::{
-    http::{
-        header::{self, HeaderValue},
-        Method as ActixMethod,
-    },
-    Error as ActixError, HttpMessage, HttpResponse,
-};
-use futures::{Future, IntoFuture, Poll};
-
-use crate::rest_api::ErrorResponse;
-#[cfg(feature = "authorization")]
-use crate::rest_api::Method;
-
-#[cfg(feature = "authorization")]
-use super::authorization::{AuthorizationHandler, PermissionMap};
-use super::{authorize, identity::IdentityProvider, AuthorizationResult};
-
-/// Authorization middleware for the Actix REST API
-pub struct AuthorizationMiddleware<S> {
-    identity_providers: Vec<Box<dyn IdentityProvider>>,
-    #[cfg(feature = "authorization")]
-    authorization_handlers: Vec<Box<dyn AuthorizationHandler>>,
-    service: S,
-}
-
-impl<S, B> Service for AuthorizationMiddleware<S>
-where
-    S: Service<Request = ServiceRequest, Response = ServiceResponse<B>, Error = ActixError>,
-    S::Future: 'static,
-    B: 'static,
-{
-    type Request = ServiceRequest;
-    type Response = ServiceResponse<B>;
-    type Error = S::Error;
-    type Future = Box<dyn Future<Item = Self::Response, Error = Self::Error>>;
-
-    fn poll_ready(&mut self) -> Poll<(), Self::Error> {
-        self.service.poll_ready()
-    }
-
-    fn call(&mut self, req: ServiceRequest) -> Self::Future {
-        if req.method() == ActixMethod::OPTIONS {
-            return Box::new(self.service.call(req).and_then(|mut res| {
-                res.headers_mut().insert(
-                    header::ACCESS_CONTROL_ALLOW_CREDENTIALS,
-                    HeaderValue::from_static("true"),
-                );
-
-                res
-            }));
-        }
-
-        #[cfg(feature = "authorization")]
-        let method = match *req.method() {
-            ActixMethod::GET => Method::Get,
-            ActixMethod::POST => Method::Post,
-            ActixMethod::PUT => Method::Put,
-            ActixMethod::PATCH => Method::Patch,
-            ActixMethod::DELETE => Method::Delete,
-            ActixMethod::HEAD => Method::Head,
-            _ => {
-                return Box::new(
-                    req.into_response(
-                        HttpResponse::BadRequest()
-                            .json(ErrorResponse::bad_request(
-                                "HTTP method not supported by Splinter REST API",
-                            ))
-                            .into_body(),
-                    )
-                    .into_future(),
-                )
-            }
-        };
-
-        let auth_header =
-            match req
-                .headers()
-                .get("Authorization")
-                .map(|auth| auth.to_str())
-                .transpose()
-            {
-                Ok(opt) => opt,
-                // Not including the error since it could leak secrets from the Authorization header
-                Err(_) => return Box::new(
-                    req.into_response(
-                        HttpResponse::BadRequest()
-                            .json(ErrorResponse::bad_request(
-                                "Authorization header must contain only visible ASCII characters",
-                            ))
-                            .into_body(),
-                    )
-                    .into_future(),
-                ),
-            };
-
-        #[cfg(feature = "authorization")]
-        let permission_map = match req.app_data::<PermissionMap<Method>>() {
-            Some(map) => map,
-            None => {
-                error!("Missing REST API permission map");
-                return Box::new(
-                    req.into_response(
-                        HttpResponse::InternalServerError()
-                            .json(ErrorResponse::internal_error())
-                            .into_body(),
-                    )
-                    .into_future(),
-                );
-            }
-        };
-
-        match authorize(
-            #[cfg(feature = "authorization")]
-            &method,
-            req.path(),
-            auth_header,
-            #[cfg(feature = "authorization")]
-            permission_map.get_ref(),
-            &self.identity_providers,
-            #[cfg(feature = "authorization")]
-            &self.authorization_handlers,
-        ) {
-            AuthorizationResult::Authorized(identity) => {
-                debug!("Authenticated user {:?}", identity);
-                req.extensions_mut().insert(identity);
-            }
-            #[cfg(any(
-                feature = "authorization",
-                feature = "biome-credentials",
-                feature = "oauth"
-            ))]
-            AuthorizationResult::NoAuthorizationNecessary => {}
-            AuthorizationResult::Unauthorized => {
-                return Box::new(
-                    req.into_response(
-                        HttpResponse::Unauthorized()
-                            .json(ErrorResponse::unauthorized())
-                            .into_body(),
-                    )
-                    .into_future(),
-                )
-            }
-            #[cfg(feature = "authorization")]
-            AuthorizationResult::UnknownEndpoint => {
-                return Box::new(
-                    req.into_response(
-                        HttpResponse::NotFound()
-                            .json(ErrorResponse::not_found("endpoint not found"))
-                            .into_body(),
-                    )
-                    .into_future(),
-                )
-            }
-        }
-
-        Box::new(self.service.call(req).and_then(|mut res| {
-            res.headers_mut().insert(
-                header::ACCESS_CONTROL_ALLOW_CREDENTIALS,
-                HeaderValue::from_static("true"),
-            );
-
-            res
-        }))
-    }
-}
 
 #[cfg(test)]
 mod tests {
     use super::*;
 
+    use actix_web::dev::*;
     use actix_web::{http::StatusCode, test, web, App, HttpRequest};
+    use actix_web::{
+        http::{header::HeaderValue, Method as ActixMethod},
+        HttpResponse,
+    };
 
     use crate::error::InternalError;
     #[cfg(feature = "authorization")]
     use crate::rest_api::auth::authorization::Permission;
+    #[cfg(feature = "authorization")]
+    use crate::rest_api::auth::authorization::PermissionMap;
+    use crate::rest_api::auth::identity::IdentityProvider;
     use crate::rest_api::auth::{identity::Identity, AuthorizationHeader};
+    #[cfg(feature = "authorization")]
+    use crate::rest_api::Method;
 
     /// Verifies that the authorization middleware sets the `Access-Control-Allow-Credentials: true`
     /// header for `OPTIONS` requests.
